todobackend/requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2020.6.20
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
jws==0.1.3
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.5
Pyrebase4==4.5.0
python-firebase==1.2
python-jwt==2.0.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wincertstore==0.2
